Move PendingCall to a new file
